+++ packages/components/CHANGELOG.md
@@ -29,6 +29,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - `Flyout Multilevel`: Secondary scroll area not shown in iOS Safari (iPhone only)
   ([3663](https://github.com/porsche-design-system/porsche-design-system/pull/3663))
 
+#### Changed
+
+- `Select`: Slotted image of selected option now visible in combobox
+  ([#3651](https://github.com/porsche-design-system/porsche-design-system/pull/3651))
+
 ### [3.22.0] - 2024-12-06
 
 ### [3.22.0-rc.1] - 2024-12-06